# MultiLineString/MultiLineString intersects performance improvements