There were two main places where memory was being allocated:

1. In the `a.Intersects(b)` call, where both `a` and `b` where `Line`s. Because both `a` and `b` get stored inside a `Geometry` interface, they get allocated to the heap rather than the stack.
2. In each stack push/pop operation, a `Line` gets stored as an `interface{}` which causes it to get stored on the heap rather than the stack.

I solved 1 by calling `intersectsLineWithLine` directly.

I solved 2 by implementing a custom heap operation which doesn't use `interface{}`.

Performance deltas are:

```
benchmark                                               old ns/op     new ns/op     delta
BenchmarkIntersectsLineStringWithLineString/n=10        7504          2186          -70.87%
BenchmarkIntersectsLineStringWithLineString/n=100       84679         21512         -74.60%
BenchmarkIntersectsLineStringWithLineString/n=1000      898498        267713        -70.20%
BenchmarkIntersectsLineStringWithLineString/n=10000     11004875      3144332       -71.43%

benchmark                                               old allocs     new allocs     delta
BenchmarkIntersectsLineStringWithLineString/n=10        128            18             -85.94%
BenchmarkIntersectsLineStringWithLineString/n=100       1214           18             -98.52%
BenchmarkIntersectsLineStringWithLineString/n=1000      12020          18             -99.85%
BenchmarkIntersectsLineStringWithLineString/n=10000     120040         18             -99.99%

benchmark                                               old bytes     new bytes     delta
BenchmarkIntersectsLineStringWithLineString/n=10        5984          1184          -80.21%
BenchmarkIntersectsLineStringWithLineString/n=100       54880         7008          -87.23%
BenchmarkIntersectsLineStringWithLineString/n=1000      515168        66144         -87.16%
BenchmarkIntersectsLineStringWithLineString/n=10000     7362656       655968        -91.09%
```

- Throughput operation speed has been reduced by ~70% (~3x speedup). This reduction would be due to decrease in memory management overhead (allocations and GC).
- The _number_ of allocations is now constant rather than linear.
- The _total amount_ allocated is still linear (although has a much smaller constant factor).